In the controller adding the functionality of the AgregarLibros(Add books), generating an auto ID for each new book that the user wants to create also showing the required message for the NombreLibro(Book's name) and AutorLibro(book's writer).
I created a Helpers folder to create a new class that has the responsability of monitoring the latest book ID of the list, this with the purpose of show it into the view of AgregarLibros(Add books).

In the model Libros, I removed the required for the bool Disponible(Book availabity) and add the nullable to the ParteAEditar and NuevoValor because those properties are only use in the EditarLibros(Book's editor).

In the views I made some adjustments to show errors and to show the info that the user prompt.